+++ projects/angular-material-extended/src/lib/matx-textarea/matx-textarea.module.ts
@@ -1,0 +1,19 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { MatxTextareaComponent } from './matx-textarea.component';
+import { MatFormFieldModule, MatInputModule } from '@angular/material';
+import { ReactiveFormsModule } from '@angular/forms';
+
+@NgModule({
+  declarations: [MatxTextareaComponent],
+  exports: [
+    MatxTextareaComponent
+  ],
+  imports: [
+    CommonModule,
+    MatFormFieldModule,
+    MatInputModule,
+    ReactiveFormsModule
+  ]
+})
+export class MatxTextareaModule { }